Reduce complexity in rasterize and transform methods. Move tiles and pipes into subpackages and break into multiple modules.
