+++ gfw_pixetl/pipes/__init__.py
@@ -1,0 +1,4 @@
+from gfw_pixetl.pipes.pipe import Pipe
+from gfw_pixetl.pipes.raster_pipe import RasterPipe
+from gfw_pixetl.pipes.raster_calc_pipe import CalcRasterPipe
+from gfw_pixetl.pipes.vector_pipe import VectorPipe
