+++ gfw_pixetl/tiles/__init__.py
@@ -1,0 +1,3 @@
+from gfw_pixetl.tiles.tile import Tile
+from gfw_pixetl.tiles.raster_src_tile import RasterSrcTile
+from gfw_pixetl.tiles.vector_src_tile import VectorSrcTile
